Move _snapshot_format.py to utils
